requirements.txt
pygame
paho-mqtt
pyserial
cairosvg